@@ -6,14 +6,13 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_rating_bar: ^4.0.0
 
-  cupertino_icons: ^1.0.0
   flutter_svg: ^0.22.0
   wechat_camera_picker: ^2.4.0
   wechat_assets_picker: ^6.0.6
