@@ -20,13 +20,13 @@ class MyApp extends StatelessWidget {
           bodyText2: TextStyle(color: AppColor.kPrimaryTextColor),
         ),
       ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
+      home: PostScreen(),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
   final String title;
 
   @override
