@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:dish/configs/constant_colors.dart';
 
 class SimpleDivider extends StatelessWidget {
-  const SimpleDivider({Key key}) : super(key: key);
+  const SimpleDivider({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
